+++ lib/pack_core/navigation/app_router.dart
@@ -3,17 +3,27 @@ import 'package:wbw_core/wbw_core.dart';
 /// !If you want to add new route, please add the route to the [routes]
 class NavigationRoutes {
   NavigationRoutes._();
-  static const mainMenu = '/main-menu';
-  static const game = '/game';
-  static const routes = [
-    mainMenu,
-    game,
-  ];
+  static const root = '/';
+  static const pause = '/pause';
+  static const allLevels = '/lvl/all';
+
+  static const levelRoot = '/lvl';
+
+  static const levelOptions = '/lvl/:id/options';
+  static String getLevelOptions({required final LevelModelId id}) =>
+      '/lvl/$id/options';
+
+  static const routes = [pause, root, levelOptions, allLevels];
 }
 
 class AppRouterController extends RouterController {
   AppRouterController.use(super.read) : super.use();
 
-  void toMainMenu() => to(NavigationRoutes.mainMenu);
-  void toGame() => to(NavigationRoutes.game);
+  void toMainMenu() => to(NavigationRoutes.pause);
+  void toGame() => to(NavigationRoutes.root);
+  void toLevelOptions({
+    required final LevelModelId id,
+  }) =>
+      to(NavigationRoutes.getLevelOptions(id: id));
+  void toAllLevel() => to(NavigationRoutes.allLevels);
 }