+++ lib/pack_game/controls/widgets/player_swither_bar.dart
@@ -3,9 +3,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:word_by_word_game/pack_core/global_states/ephemeral/ephemeral.dart';
 
-// TODO(arenukvern): listen  diDto.levelPlayersBloc changes
-/// (changed from empty or loading to the live state)
-///
 class PlayerSwitcher extends HookWidget {
   const PlayerSwitcher({super.key});