+++ lib/pack_game/levels/screens/all_levels_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/src/widgets/container.dart';
+import 'package:flutter/src/widgets/framework.dart';
+
+part 'levels_chooser_screen_state.dart';
+
+class AllLevelsScreen extends StatelessWidget {
+  const AllLevelsScreen({super.key});
+
+  @override
+  Widget build(final BuildContext context) {
+    return Container();
+  }
+}