+++ lib/pack_game/levels/screens/level_options_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/src/widgets/container.dart';
+import 'package:flutter/src/widgets/framework.dart';
+
+part 'level_options_screen_state.dart';
+
+class LevelOptionsScreen extends StatelessWidget {
+  const LevelOptionsScreen({super.key});
+
+  @override
+  Widget build(final BuildContext context) {
+    return Container();
+  }
+}