+++ lib/pack_game/main_menu/main_menu.dart
@@ -1,1 +1,0 @@
-export 'main_menu_screen.dart';