+++ lib/pack_game/pack_game.dart
@@ -1,5 +1,5 @@
 export 'game/game.dart';
 export 'levels/levels.dart';
 export 'levels_hud/levels_hud.dart';
-export 'main_menu/main_menu.dart';
 export 'mechanics/mechanics.dart';
+export 'pause/pause.dart';