+++ lib/pack_game/pause/pause.dart
@@ -1,0 +1,1 @@
+export 'pause_screen.dart';